pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbreaker"
version = "0.1.0"
description = "Autonomous AI red-teaming that finds what manual testing can't."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "openai>=1.30.0",
    "anthropic>=0.25.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "pillow>=10.3.0",
    "tenacity>=8.3.0",
]

[project.scripts]
agentbreaker = "agentbreaker.cli:main"

[tool.setuptools.packages.find]
include = ["agentbreaker*"]
